createTicket function

+++ domain/flightInfo/Ticket.java
@@ -2,13 +2,13 @@ package flightInfo;
 import util.*;
 public class Ticket {
     private int seatNum;
-    private int price;
+    private double price;
     private String departureTime;
     private boolean hasCancellationInsurance = false;
     private String classSeat;
 
     // Constructor ONLY CALLED BY CONTROLLER
-    public Ticket(int seatNum, int price, boolean hasCancellationInsurance, Name name,String departureTime, String classSeat) {
+    public Ticket(int seatNum, double price, boolean hasCancellationInsurance, Name name,String departureTime, String classSeat) {
         this.seatNum = seatNum;
         this.price = price;
         this.hasCancellationInsurance = hasCancellationInsurance;
@@ -21,12 +21,12 @@ public class Ticket {
     }
 
     // Getter for price
-    public int getPrice() {
+    public double getPrice() {
         return price;
     }
 
     // Setter for price
-    public void setPrice(int price) {
+    public void setPrice(double price) {
         this.price = price;
     }
